more mvp refactor
