+++ app/src/main/java/net/rmoreno/weatherapp/DailyWeatherPresenterImpl.java
@@ -1,34 +1,23 @@
 package net.rmoreno.weatherapp;
 
-import android.content.Intent;
 import android.util.Log;
-import android.view.View;
 
-import com.squareup.okhttp.Call;
-import com.squareup.okhttp.Callback;
-import com.squareup.okhttp.OkHttpClient;
-import com.squareup.okhttp.Request;
-import com.squareup.okhttp.Response;
-
-import net.rmoreno.weatherapp.models.CurrentWeather;
 import net.rmoreno.weatherapp.models.DailyWeather;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.io.IOException;
 import java.util.ArrayList;
 
 public class DailyWeatherPresenterImpl implements DailyWeatherPresenter {
 
-    WeatherNetwork network;
     DailyWeatherView view;
 
-    DailyWeatherPresenterImpl(DailyWeatherView view, WeatherNetwork network) {
-        this.network = network;
+    DailyWeatherPresenterImpl(DailyWeatherView view) {
         this.view = view;
     }
+
     @Override
     public void resume() {
 
@@ -46,50 +35,11 @@ public class DailyWeatherPresenterImpl implements DailyWeatherPresenter {
 
     @Override
     public DailyWeather getDailyWeather(float lat, float lng) {
-        network.getDailyWeather(lat, lng, new Callback() {
-            @Override
-            public void onFailure(Request request, IOException e) {
 
-            }
-
-            @Override
-            public void onResponse(Response response) throws IOException {
-                try{
-                    String jsonData = response.body().string();
-                    if (response.isSuccessful()) {
-                        CurrentWeather currentWeather = getCurrentWeatherData(jsonData);
-                        ArrayList<DailyWeather> dailyWeather = getDailyWeatherData(jsonData);
-                        view.displayDailyWeather();
-                    }
-                }  catch(JSONException e) {
-                    Log.d("DWPresen" + " JSONEXCEPTION", e.getMessage());
-                } catch(IOException e){
-                    Log.d("DWPresent" + " IOEXCEPTION", e.getMessage());
-                }
-
-            }
-        });
         return new DailyWeather();
     }
 
-    public CurrentWeather getCurrentWeatherData(String jsonData) throws JSONException{
 
-        CurrentWeather currentWeather = new CurrentWeather();
-        JSONObject jsonObject = new JSONObject(jsonData);
-
-        JSONObject currently = jsonObject.getJSONObject("currently");
-
-        currentWeather.setTemp(currently.getDouble("temperature"));
-        currentWeather.setSummary(currently.getString("summary"));
-        currentWeather.setTime(currently.getLong("time"));
-        currentWeather.setTimeZone(jsonObject.getString("timezone"));
-        currentWeather.setIcon(currently.getString("icon"));
-        currentWeather.setPrecip(currently.getInt("precipProbability"));
-        currentWeather.setFeels(currently.getDouble(("apparentTemperature")));
-        currentWeather.setWind(currently.getDouble(("windSpeed")));
-
-        return currentWeather;
-    }
 
     public ArrayList<DailyWeather> getDailyWeatherData(String jsonData) throws JSONException {
         ArrayList<DailyWeather> dailyWeatherList = new ArrayList<>();
